refactor(systems): 装飾の導出を render3d から decoration.go へ分離する
収納マーカーの「何を出すか」の導出を render3d.go から decoration.go へ切り出し、
render3d は渡された装飾クアッドを描くだけにする。マーカーは状態従属・視点相対の
ヒントなのでエンティティ化せず導出のまま。血痕・破片・撃破のようなイベント駆動で
残す装飾は別途エンティティを spawn し collectBillboards/VisualEffectSystem が描く。
appendMarkerQuad を高さと tint を引数に取る汎用 appendTileBillboard へ一般化し、
導出系の装飾が升にカメラ正面の板を立てる幾何を共有できるようにする。
